# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "School Year API",
        "description": "API endpoints for managing school years",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/add-school-year": {
            "post": {
                "tags": [
                    "School Years"
                ],
                "summary": "Add a new school year",
                "description": "Creates a new school year with grade level and section",
                "operationId": "storeSchoolYear",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "school_year",
                                    "grade_level",
                                    "section_name"
                                ],
                                "properties": {
                                    "school_year": {
                                        "type": "string",
                                        "example": "2025-2026"
                                    },
                                    "grade_level": {
                                        "type": "string",
                                        "example": "Grade 1"
                                    },
                                    "section_name": {
                                        "type": "string",
                                        "example": "Section A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "School year created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "School year, grade level, and section added successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "school_year": {
                                                    "type": "string",
                                                    "example": "2025-2026"
                                                },
                                                "grade_level": {
                                                    "type": "string",
                                                    "example": "Grade 1"
                                                },
                                                "section_name": {
                                                    "type": "string",
                                                    "example": "Section A"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "datetime"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/school-years": {
            "get": {
                "tags": [
                    "School Years"
                ],
                "summary": "Get all school years",
                "description": "Returns a list of all school years",
                "operationId": "getSchoolYears",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "School years retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "school_year": {
                                                        "type": "string",
                                                        "example": "2025-2026"
                                                    },
                                                    "grade_level": {
                                                        "type": "string",
                                                        "example": "Grade 1"
                                                    },
                                                    "section_name": {
                                                        "type": "string",
                                                        "example": "Section A"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "School Years",
            "description": "API Endpoints for School Year operations"
        }
    ]
}